+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Fixed
 
 - Fix Lua bind of static methods for remote machines
 - Fix rollup-memory-range decoding of empty payloads at end of file
+
+### Changed
+
+- Improve server-manager concurrency violation messages
 
 ## [0.10.0] - 2022-07-04
 ### Added